fix: unsubscribe method

+++ src/main/resources/db/migration/V1.1__create_subscriber_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE subscriber (
-	id int8 NOT NULL,
+	id SERIAL NOT NULL,
 	"name" varchar(255) NULL,
 	"surname" varchar(255) NULL,
     "email" varchar(255) NULL,
